Backend del segundo reporte funcionando con parametros

+++ tallerMecanicoBackEnd/src/main/java/com/TallerMecanicoSpring/TM/model/RsReporteCantServMarcaEntreFecha.java
@@ -3,7 +3,7 @@ package com.TallerMecanicoSpring.TM.model;
 public class RsReporteCantServMarcaEntreFecha {
    private String marca;
    private String nombreServicio;
-   private Integer cantidad;
+   private Long cantidad;
 
    public String getMarca() {
     return marca;
@@ -17,10 +17,10 @@ public class RsReporteCantServMarcaEntreFecha {
     public void setNombreServicio(String nombreServicio) {
         this.nombreServicio = nombreServicio;
     }
-    public Integer getCantidad() {
+    public Long getCantidad() {
         return cantidad;
     }
-    public void setCantidad(Integer cantidad) {
+    public void setCantidad(Long cantidad) {
         this.cantidad = cantidad;
     } 
 }
